requirements.txt
python-telegram-bot[job-queue]
python-dotenv
pymorphy2